Función terminada

+++ database/migrations/2020_08_27_155456_table_product_cars_create.php
@@ -5,9 +5,9 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 class TableProductCarsCreate extends Migration{
-    public function up(){
-        Schema::create('product_cars', function (Blueprint $table) {
-            $table->bigIncrements('product_id');
+   public function up(){
+        Schema::create('product_cars', function (Blueprint $table){
+            $table->integer('product_id');
             $table->integer('cart_id');
             $table->integer('quantity')->default(1);
         });
